requirements.txt
flask
requests
readability
readability-lxml
html2text
azure-cosmos
urllib3
azure-ai-textanalytics
azure-core